requirements.txt
alembic==1.8.1
celery==5.0.0
fastapi==0.79.0
gunicorn==20.1.0
pydantic==1.9.2
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
python-decouple==3.6
uvicorn==0.18.2
psycopg2==2.9.1
pythondi==1.2.4